Get Tests to compile on Linux
